krssh: call tryParse on fewer packets
We're looking for specific packets that happen early in a SSH connection.

Once we send the host auth packet to krd, we should not continue trying to
parse every packet coming back from the server.

I agree to license all rights to my contributions in each modified file
exclusively to KryptCo, Inc.

Signed-off-by: tcely <tcely@users.noreply.github.com>